+++ gruntfile.js
@@ -3,29 +3,12 @@ module.exports = function (grunt) {
     grunt.loadTasks('tasks');
 
     grunt.registerTask('test', 'Run tests against raw source', [
-        'jasmine:src',
-        'jshint'
+        'jasmine'
     ]);
-    grunt.registerTask('default', 'Build suite.', [
-        'jasmine:src',
+    grunt.registerTask('default', 'Full build suite.', [
+        'jasmine',
         'jshint',
-        'uglify',
-        'jasmine:dist-global',
-        'jasmine:dist-single',
-        'jasmine:dist-other'
-    ]);
-    grunt.registerTask('dev-test', 'Run SpecRunner.html locally.', [
-        'jasmine:dist-global',
-        'jasmine:dist-single',
-        'jasmine:dist-other',
-        'connect',
-        'watch'
-    ]);
-    grunt.registerTask('sl-test', 'SauceLabs test suite.', [
-        'jasmine:dist-global',
-        'jasmine:dist-single',
-        'jasmine:dist-other',
-        'connect',
-        'saucelabs-jasmine:all'
+        'browserify',
+        'uglify'
     ]);
 };